fix abomination
I've recently learned that in C, that a function with no parameters
should be declared with 'void' in the argument list. Leaving empty
parens implies that the parameters are unspecified, and therefore the
arguments can in fact be called with one or more arguments.

This is in contrast with C++ where leaving empty parens signifies that
the function takes no arguments, and it is syntactically incorrect to do
otherwise.

tl;dr

C  : void foo(void);
C++: void foo();

http://en.wikipedia.org/wiki/Compatibility_of_C_and_C%2B%2B