Fix off-by-one bug in response callback
I wasn't aware of some of the subtleties of strncat when I was
originally writing this, like how it automatically appends an additional
NULL byte onto the destination string. I hadn't observed any ill effects
of this previously, but it's safer to fix it now.

Also, rename `callback()' to `resp_callback()' and fix tests for this
commit.