Change type of *hdrc variables to int
size_t's should only be used for the size of object in bytes, not just
for any general size of something. in this situation, an unsigned int is
probably the most appropriate, however this can potentially cause bugs
down the line with mixed sign arithmetic, so we'll just use ints.

also, argc is an int ;)